Initial commit for the Itus Network Shield specific changes to the OpenWrt Source.
This commit includes some utility scripts I use so I don't have to remember
commands.  Yes, I'm lazy like that.

okernel - Open the kernel for editing/patching
ckernel - Close the kernel for building
xferimage - Transfer the image to the Shield
            This script ignores ssh fingerprints, as they change after each flash.
            Script can be edited to change the image you are working on.

Commit includes the custom /files to be included/altered for the install.